requirements.txt
numpy
matplotlib
scikit-learn
streamlit
flask
langchain
langchain_community
langchain_core
langchain-huggingface
langserve
uvicorn
pypdf
faiss-cpu
PyPDF2
tiktoken
langchain-google-genai
python-dotenv
pytest

